requirements.txt
opencv-python
mediapipe
numpy
pandas
pyttsx3
luma.oled
Pillow
picamera2
